external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/implicit/assign_ivec4_to_vec4.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:4: '=' :  cannot convert from 'const mediump 4-component vector of int' to '4-component vector of float'

implicit cast assigning ivec4 to vec4 should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: implicit cast assigning ivec4 to vec4 should fail
test vertex shader
default fragment shaderPASS implicit cast assigning ivec4 to vec4 should fail
PASS successfullyParsed is true

TEST COMPLETE
